moved API to new endpoint location
