+++ app/api/api_v1/api.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from .endpoints import login, user, data, rule, unit, operator, tool, pest_model, parcel, disease
+from .endpoints import login, user, data, rule, unit, operator, tool, pest_model, parcel, disease, model
 
 api_router = APIRouter()
 api_router.include_router(login.router, prefix="/login", tags=["login"])
@@ -12,3 +12,4 @@ api_router.include_router(tool.router, prefix="/tool", tags=["tool"])
 api_router.include_router(pest_model.router, prefix="/pest-model", tags=["pest-model"])
 api_router.include_router(parcel.router, prefix="/parcel", tags=["parcel"])
 api_router.include_router(disease.router, prefix="/disease", tags=["disease"])
+api_router.include_router(model.router, prefix="/model", tags=["model"])
