+++ app/init/init_gatekeeper.py
@@ -5,7 +5,8 @@ from core import settings
 from requests.exceptions import RequestException
 
 
-from api.api_v1.endpoints import operator, pest_model, rule, tool, unit
+from api.api_v1.endpoints import operator, pest_model, rule, tool, unit, model, disease
+
 
 def register_apis_to_gatekeeper():
 
@@ -34,6 +35,8 @@ def register_apis_to_gatekeeper():
     apis_to_register.include_router(rule.router, prefix="/rule")
     apis_to_register.include_router(tool.router, prefix="/tool")
     apis_to_register.include_router(unit.router, prefix="/unit")
+    apis_to_register.include_router(model.router, prefix="/model")
+    apis_to_register.include_router(disease.router, prefix="/disease")
 
     for api in apis_to_register.routes:
         try:
@@ -43,7 +46,7 @@ def register_apis_to_gatekeeper():
                 json={
                     "base_url": "http://{}:{}/".format(settings.SERVICE_NAME, settings.SERVICE_PORT),
                     "service_name": settings.SERVICE_NAME,
-                    "endpoint": "api/v1/" + api.path.strip("/"),
+                    "endpoint": "api/v1/" + api.path.lstrip("/"),
                     "methods": list(api.methods)
                 }
             )
